NLMPPI-37 | refactor adapter test for CartPole-v1
